requirements.txt
annotated-types==0.7.0
anyio==4.11.0
av==15.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.2
ctranslate2==4.6.0
Deprecated==1.2.18
deprecation==2.1.0
fastapi==0.118.2
faster-whisper==1.2.0
gunicorn>=21.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
limits==5.6.0
llvmlite==0.44.0
lxml==6.0.2
MarkupSafe==3.0.2
memory-profiler==0.61.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
openai-whisper==20250625
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
postgrest==2.22.0
propcache==0.4.1
protobuf==6.32.0
psutil==5.9.6
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.1.1
pdfminer.six==20231228
pdf2image==1.17.0
pytesseract==0.3.13
pyreadline3==3.5.4
pytest==8.4.2
pytest-asyncio==1.2.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.22.0
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
